Compile workaround for Swift 3.2 and Swift 4

+++ Sources/MongoKitten/CRUD/CRUD.swift
@@ -273,11 +273,11 @@ extension CollectionQueryable {
         }
         
         if let skip = skip {
-            command["skip"] = Int32(skip)
+            command["skip"] = Int32(skip) as Int32
         }
         
         if let limit = limit {
-            command["limit"] = Int32(limit)
+            command["limit"] = Int32(limit) as Int32
         }
         
         command["readConcern"] = readConcern ?? self.readConcern
@@ -508,11 +508,11 @@ extension CollectionQueryable {
             }
             
             if let skip = skip {
-                command["skip"] = Int32(skip)
+                command["skip"] = Int32(skip) as Int32
             }
             
             if let limit = limit {
-                command["limit"] = Int32(limit)
+                command["limit"] = Int32(limit) as Int32
             }
             
             if let listener = database.server.whenExplaining {
